requirements.txt
fastapi
uvicorn
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pandas
openpyxl
python-dotenv
schedule
google-generativeai